First pass at adding support for streaming jobs to pycascading
Example files are in examples/streaming_job.py and examples/streaming_task.py, where the
latter is the python run externally.

so far I've added support for stream_add and stream_map_to. All the outputs from the stream are treated as strings, though you can name them.

A thread is ran for the purposes of reading in the stream and periodically dumped to the output flow.

Far more comments and some tests are needed.